organizing code and folders
