+++ controllers/profile.js
@@ -1,0 +1,21 @@
+const handleProfileGet = (req, res, db) => {
+	const { id } = req.params;
+	db.select('*').from('users').where({
+		id: id
+			})
+		.then(user =>{
+			if (user.length) {
+				res.json(user[0])
+			} else {
+				res.status(400).json('user not found sorry!')
+			}
+		
+		})
+	.catch(err => res.status(400).json('bullshit error after searching for user'))
+}
+
+
+module.exports = {
+	handleProfileGet: handleProfileGet
+}
+
