+++ controllers/signin.js
@@ -1,0 +1,23 @@
+const handleSignin = (db, bcrypt )=>(req, res) => {
+db.select('email', 'hash').from('login')
+	.where('email', '=', req.body.email)
+	.then(data => {
+	const isValid =	bcrypt.compareSync(req.body.password, data[0].hash);
+		if (isValid) {
+			return db.select('*').from('users')
+			.where('email', '=', req.body.email)
+			.then(user => {
+				res.json(user[0])
+			})
+			.catch(err=> res.status(400).json('User does not exist'))
+		} else {
+
+		res.status(400).json('wrong credentials friend')
+	}
+	})
+	.catch(err=> res.status(400).json('error at sign in route'))
+}
+
+module.exports = {
+	handleSignin: handleSignin
+}
