+++ server.js
@@ -2,9 +2,10 @@ const express = require('express');
 const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
 const knex = require('knex');
-
-
-
+const register = require('./controllers/register');
+const signin = require('./controllers/signin');
+const profile = require('./controllers/profile');
+const image	= require('./controllers/image');	
 
 const db = knex({
   client: 'pg',
@@ -34,81 +35,13 @@ app.get('/', (req, res) => {
 })
 
 
-app.post('/signin', (req, res) => {
-db.select('email', 'hash').from('login')
-	.where('email', '=', req.body.email)
-	.then(data => {
-	const isValid =	bcrypt.compareSync(req.body.password, data[0].hash);
-		if (isValid) {
-			return db.select('*').from('users')
-			.where('email', '=', req.body.email)
-			.then(user => {
-				res.json(user[0])
-			})
-			.catch(err=> res.status(400).json('User does not exist'))
-		} else {
+app.post('/signin', (req, res) => signin.handleSignin(db, bcrypt))
 
-		res.status(400).json('wrong credentials friend')
-	}
-	})
-	.catch(err=> res.status(400).json('error at sign in route'))
-})
+app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) })
 
-app.post('/register', (req, res) => {
-	const { email, name, password } = req.body;
-	const hash = bcrypt.hashSync(password);
-	db.transaction(trx=> {
-		trx.insert({
-			hash: hash,
-			email: email
-		})
-		.into('login')
-		.returning('email')
-		.then(loginEmail=>{
-			return trx('users')
-			.returning('*')
-			.insert({
-			email: loginEmail[0],
-			name: name,
-			joined: new Date()
-			})
-			.then(user => {
-			res.json(user[0]);
-		})
-		})
-		.then(trx.commit)
-		.catch(trx.rollback)
-	})	
-		.catch(err => res.status(400).json('sorry unable to register'))
-	})
-
-app.get('/profile/:id', (req, res) => {
-	const { id } = req.params;
-	db.select('*').from('users').where({
-		id: id
-			})
-		.then(user =>{
-			if (user.length) {
-				res.json(user[0])
-			} else {
-				res.status(400).json('user not found sorry!')
-			}
+app.get('/profile/:id', (req, res,) => {profile.handleProfileGet(res, req, db)})
 		
-	})
-		.catch(err => res.status(400).json('bullshit error after searching for user'))
-})
-
-app.put('/image', (req, res) => {
-	const { id } = req.body;
-	 db('users').where('id', '=', id)
-  	.increment('entries', 1)
-  	.returning('entries')
-  	.then(entries=>
-  	{
-  		res.json(entries[0]);
-  	})
-	.catch(err => res.status(400).jason('unable to get entries'))
-})
+app.put('/image', (req, res) => {image.imageHandler(req, res, db)})
 
 // bcrypt.hash("bacon", null, null, function(err, hash) {
 //     // Store hash in your password DB.
